refactor: modularize example plugin and adjust docs/training material to align with v4 BaseTool pattern
